refactor: refactore code and also change folder structure
