+++ app/(app)/home/Hero.tsx
