+++ app/(app)/home/page.tsx
@@ -1,0 +1,9 @@
+import Home from "./Home";
+
+export default function HomePage() {
+  return (
+    <>
+      <Home />
+    </>
+  );
+}
