+++ app/(app)/page.tsx
@@ -1,9 +1,6 @@
+import { redirect } from "next/navigation";
 import Home from "./home/Home";
 
 export default async function Page() {
-  return (
-    <>
-      <Home />
-    </>
-  );
+  return <Home />;
 }
