+++ app/(auth)/layout.tsx
@@ -1,0 +1,11 @@
+export default function AuthLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <main className="flex min-h-screen items-center justify-center bg-muted">
+      <div className="w-full max-w-sm px-4">{children}</div>
+    </main>
+  );
+}
