+++ app/(auth)/signin/page.tsx
@@ -1,0 +1,16 @@
+import { redirect } from "next/navigation";
+import Signin from "@/components/common/Signin";
+import { getAuthSession } from "@/lib/service/auth";
+
+export default async function SigninPage() {
+  const session = await getAuthSession();
+
+  if (session) {
+    redirect("/");
+  }
+  return (
+    <>
+      <Signin />
+    </>
+  );
+}
